:hammer: changed submenu with hover menu
